pyproject.toml
[tool.poetry]
name = "verse-inserter"
version = "1.0.0"
description = "Automated scripture insertion for Microsoft Word documents"
authors = ["Kasim Lyee <lyee@codewithlyee.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://codewithlyee.com/stores/verse-inserter"
repository = "https://github.com/kasimlyee/verse-inserter"
documentation = "https://docs.codewithlyee.com/verse-inserter"
keywords = ["bible", "scripture", "word", "automation", "docx"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Religion",
    "Topic :: Text Processing :: Markup",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.12",
]
include = ["LICENSE", "README.md", "verse_inserter/ui/resources/**/*"]

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
python-docx = "^1.1.2"
requests = "^2.31.0"
cryptography = "^42.0.0"
ttkbootstrap = "^1.10.1"
pydantic = "^2.6.0"
pydantic-settings = "^2.2.0"
tenacity = "^8.2.3"
aiohttp = "^3.9.0"
diskcache = "^5.6.3"
python-dotenv = "^1.0.1"
Pillow = "^10.2.0"
colorlog = "^6.8.2"
pyinstaller = "^6.16.0"
beautifulsoup4 = "^4.14.2"
lxml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = "^24.1.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pylint = "^3.0.0"
pre-commit = "^3.6.0"

[tool.poetry.scripts]
verse-inserter = "verse_inserter.__main__:main"

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers --cov=verse_inserter --cov-report=html --cov-report=term"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["verse_inserter"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
precision = 2
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"